feat: FLUI-139 Add a theme switcher, allowing switching theme from ferlab-ui to clin
