+++ storybook/Makefile
@@ -1,0 +1,9 @@
+.ONESHELL:
+
+# Work with local clin-portal-theme
+clin_theme_local:
+	npm install ../../clin-portal-theme
+
+clin_theme_external:
+	sed -i '' '/clin-portal-theme/d' ./package.json
+	npm install github:Ferlab-Ste-Justine/clin-portal-theme
